register , update ,delete

+++ src/main/java/org/dongho/board/controller/BoardController.java
@@ -1,16 +1,22 @@
 package org.dongho.board.controller;
 
+import javax.validation.Valid;
+
 import org.dongho.board.dto.BoardDTO;
 import org.dongho.board.service.BoardService;
 import org.dongho.common.dto.PageDTO;
 import org.dongho.common.dto.PageMaker;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j;
@@ -34,28 +40,53 @@ public class BoardController {
 	}
 	
 	@GetMapping({"/read" , "/modify"})
-	public void readGet(@ModelAttribute("pageDTO") PageDTO pageDTO, @ModelAttribute("bno") Integer bno ,Model model) {
+	public void readGet(PageDTO pageDTO, Integer bno ,Model model) {
 		
 		model.addAttribute("board" , service.selectOne(bno));
 		model.addAttribute("pageMaker" , new PageMaker(pageDTO , service.getTotalCount(pageDTO)));
 	}
 	
-	@GetMapping("/register")
-	public void registerGet(@ModelAttribute("pageDTO") PageDTO pageDTO , Model model) {
+	@PostMapping("/modify")
+	@ResponseBody
+	public void modifyPost(@RequestBody BoardDTO boardDTO) {
+		
+		log.info(boardDTO);
+		service.update(boardDTO);
 		
 	}
 	
-	@PostMapping("/register")
-	public void registerPost(@ModelAttribute("boardDTO") BoardDTO boardDTO,Model model) {
+	
+	
+	@GetMapping("/register")
+	public void registerGet(PageDTO pageDTO , Model model) {
+		
+	}
+	
+	@PostMapping(value = {"/register"} , produces = {"text/plain"})
+	@ResponseBody
+	public String registerPost(@RequestBody BoardDTO boardDTO) {
+		
+		log.info(boardDTO);
+		service.insert(boardDTO);
+		
+		return "success";
 		
 	}
 	
 	@PostMapping(value = {"/remove"} , produces = {"text/plain"})
-	public String removePost(@ModelAttribute("bno") Integer bno,BindingResult result,Model model) {
+	@ResponseBody
+	public ResponseEntity<String> removePost(@Valid @RequestBody Integer bno , BindingResult result ) {
+		log.info("=========================");
+		log.info(bno);
 		
+		if(result.hasErrors()) {
+			return new ResponseEntity<String>("fail" , HttpStatus.NOT_FOUND);
+		}
+		
+
 		service.delete(bno);
 		
-		return "redirect:/board/list";
+		return new ResponseEntity<String>("success",HttpStatus.OK);
 	}
 		
 }
